refactor: consolidate storages
Closes #207
