+++ cmd/kms-rest/startcmd/start.go
@@ -11,7 +11,6 @@ import (
 	"errors"
 	"fmt"
 	"net/http"
-	"strconv"
 	"strings"
 	"time"
 
@@ -39,7 +38,6 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/trustbloc/edge-core/pkg/log"
 	"github.com/trustbloc/edge-core/pkg/restapi/logspec"
-	cmdutils "github.com/trustbloc/edge-core/pkg/utils/cmd"
 	tlsutils "github.com/trustbloc/edge-core/pkg/utils/tls"
 	zcapldcore "github.com/trustbloc/edge-core/pkg/zcapld"
 
@@ -54,192 +52,8 @@ import (
 )
 
 const (
-	hostURLFlagName  = "host-url"
-	hostURLFlagUsage = "The URL to run the KMS instance on. Format: HostName:Port."
-	hostURLEnvKey    = "KMS_HOST_URL"
-
-	hostMetricsURLFlagName  = "host-metrics-url"
-	hostMetricsURLFlagUsage = "URL that exposes the metrics endpoint. Format: HostName:Port."
-	hostMetricsURLEnvKey    = "KMS_HOST_METRICS_URL"
-
-	baseURLFlagName  = "base-url"
-	baseURLEnvKey    = "KMS_BASE_URL"
-	baseURLFlagUsage = "An optional base URL value to prepend to a location returned in the Location header. " +
-		commonEnvVarUsageText + baseURLEnvKey
-
-	logLevelFlagName        = "log-level"
-	logLevelEnvKey          = "KMS_LOG_LEVEL"
-	logLevelFlagShorthand   = "l"
-	logLevelPrefixFlagUsage = "Logging level to set. Supported options: critical, error, warning, info, debug. " +
-		`Defaults to "info". ` + commonEnvVarUsageText + logLevelEnvKey
-
-	commonEnvVarUsageText = "Alternatively, this can be set with the following environment variable: "
-
-	didDomainFlagName  = "did-domain"
-	didDomainFlagUsage = "URL to the did consortium's domain." +
-		" Alternatively, this can be set with the following environment variable: " + didDomainEnvKey
-	didDomainEnvKey = "KMS_DID_DOMAIN"
-)
-
-// TLS options.
-const (
-	tlsSystemCertPoolFlagName      = "tls-systemcertpool"
-	tlsSystemCertPoolFlagShorthand = "s"
-	tlsSystemCertPoolFlagUsage     = "Use system certificate pool. Possible values [true] [false]. " +
-		"Defaults to false if not set. " + commonEnvVarUsageText + tlsSystemCertPoolEnvKey
-	tlsSystemCertPoolEnvKey = "KMS_TLS_SYSTEMCERTPOOL"
-
-	tlsCACertsFlagName      = "tls-cacerts"
-	tlsCACertsFlagShorthand = "c"
-	tlsCACertsFlagUsage     = "Comma-separated list of CA certs path. " + commonEnvVarUsageText + tlsCACertsEnvKey
-	tlsCACertsEnvKey        = "KMS_TLS_CACERTS"
-
-	tlsServeCertPathFlagName  = "tls-serve-cert"
-	tlsServeCertPathFlagUsage = "The path to the server certificate to use when serving HTTPS. " +
-		commonEnvVarUsageText + tlsServeCertPathEnvKey
-	tlsServeCertPathEnvKey = "KMS_TLS_SERVE_CERT"
-
-	tlsServeKeyPathFlagName  = "tls-serve-key"
-	tlsServeKeyPathFlagUsage = "The path to the private key to use when serving HTTPS. " +
-		commonEnvVarUsageText + tlsServeKeyPathFlagEnvKey
-	tlsServeKeyPathFlagEnvKey = "KMS_TLS_SERVE_KEY"
-)
-
-// Storage for Keystore metadata.
-const (
-	databaseTypeFlagName  = "database-type"
-	databaseTypeEnvKey    = "KMS_DATABASE_TYPE"
-	databaseTypeFlagUsage = "The type of database to use for storing metadata about keystores and " +
-		"associated keys. Supported options: mem, couchdb, mongodb. " + commonEnvVarUsageText + databaseTypeEnvKey
-
-	databaseURLFlagName  = "database-url"
-	databaseURLEnvKey    = "KMS_DATABASE_URL"
-	databaseURLFlagUsage = "The URL of the database. Not needed if using in-memory storage. " +
-		"For CouchDB, include the username:password@ text if required. " + commonEnvVarUsageText + databaseURLEnvKey
-
-	databasePrefixFlagName  = "database-prefix"
-	databasePrefixEnvKey    = "KMS_DATABASE_PREFIX"
-	databasePrefixFlagUsage = "An optional prefix to be used when creating and retrieving the underlying database. " +
-		commonEnvVarUsageText + databasePrefixEnvKey
-)
-
-// Path to the file with key to be used by secret lock to protect primary key.
-const (
-	secretLockKeyPathFlagName  = "secret-lock-key-path"
-	secretLockKeyPathEnvKey    = "KMS_SECRET_LOCK_KEY_PATH" //nolint:gosec // not hard-coded credentials
-	secretLockKeyPathFlagUsage = "The path to the file with key to be used by local secret lock. If missing noop " +
-		"service lock is used. " + commonEnvVarUsageText + secretLockKeyPathEnvKey
-)
-
-// Storage for primary keys.
-const (
-	primaryKeyDatabaseTypeFlagName  = "primary-key-database-type"
-	primaryKeyDatabaseTypeEnvKey    = "KMS_PRIMARY_KEY_DATABASE_TYPE"
-	primaryKeyDatabaseTypeFlagUsage = "The type of database to use for storing primary keys. " +
-		"Supported options: mem, couchdb, mongodb. " + commonEnvVarUsageText + primaryKeyDatabaseTypeEnvKey
-
-	primaryKeyDatabaseURLFlagName  = "primary-key-database-url"
-	primaryKeyDatabaseURLEnvKey    = "KMS_PRIMARY_KEY_DATABASE_URL"
-	primaryKeyDatabaseURLFlagUsage = "The URL of the database for primary keys. Not needed if using in-memory " +
-		"storage. For CouchDB, include the username:password@ text if required. " + commonEnvVarUsageText +
-		primaryKeyDatabaseURLEnvKey
-
-	primaryKeyDatabasePrefixFlagName  = "primary-key-database-prefix"
-	primaryKeyDatabasePrefixEnvKey    = "KMS_PRIMARY_KEY_DATABASE_PREFIX"
-	primaryKeyDatabasePrefixFlagUsage = "An optional prefix to be used when creating and retrieving the underlying " +
-		"database for primary keys. " + commonEnvVarUsageText + primaryKeyDatabasePrefixEnvKey
-)
-
-// Storage for local KMS (supports Keystore).
-const (
-	localKMSDatabaseTypeFlagName  = "local-kms-database-type"
-	localKMSDatabaseTypeEnvKey    = "KMS_LOCAL_KMS_DATABASE_TYPE"
-	localKMSDatabaseTypeFlagUsage = "The type of database to use for storing local KMS secrets (e.g. keys for " +
-		"Keystore). Supported options: mem, couchdb, mongodb. " + commonEnvVarUsageText + localKMSDatabaseTypeEnvKey
-
-	localKMSDatabaseURLFlagName  = "local-kms-database-url"
-	localKMSDatabaseURLEnvKey    = "KMS_LOCAL_KMS_DATABASE_URL"
-	localKMSDatabaseURLFlagUsage = "The URL of the database for local KMS. Not needed if using in-memory storage. " +
-		"For CouchDB, include the username:password@ text if required. " +
-		commonEnvVarUsageText + localKMSDatabaseURLEnvKey
-
-	localKMSDatabasePrefixFlagName  = "local-kms-database-prefix"
-	localKMSDatabasePrefixEnvKey    = "KMS_LOCAL_KMS_DATABASE_PREFIX"
-	localKMSDatabasePrefixFlagUsage = "An optional prefix to be used when creating and retrieving the underlying " +
-		"local KMS database. " + commonEnvVarUsageText + localKMSDatabasePrefixEnvKey
-)
-
-// Storage for Key Manager (KMS that works with user's keys).
-const (
-	keyManagerStorageTypeFlagName  = "key-manager-storage-type"
-	keyManagerStorageTypeEnvKey    = "KMS_KEY_MANAGER_STORAGE_TYPE"
-	keyManagerStorageTypeFlagUsage = "The type of storage to use for key manager. Supported options: mem, couchdb, " +
-		"mongodb, edv. " + commonEnvVarUsageText + keyManagerStorageTypeEnvKey
-
-	keyManagerStorageURLFlagName  = "key-manager-storage-url"
-	keyManagerStorageURLEnvKey    = "KMS_KEY_MANAGER_STORAGE_URL"
-	keyManagerStorageURLFlagUsage = "The URL of storage for key manager. Not needed if using in-memory storage. " +
-		"For CouchDB, include the username:password@ text if required. " + commonEnvVarUsageText +
-		keyManagerStorageURLEnvKey
-
-	keyManagerStoragePrefixFlagName  = "key-manager-storage-prefix"
-	keyManagerStoragePrefixEnvKey    = "KMS_KEY_MANAGER_STORAGE_PREFIX"
-	keyManagerStoragePrefixFlagUsage = "An optional prefix to be used when creating and retrieving the underlying " +
-		"key manager storage. " + commonEnvVarUsageText + keyManagerStoragePrefixEnvKey
-)
-
-// Cache expiration (used for EDV and Hub Auth calls).
-const (
-	cacheExpirationFlagName  = "cache-expiration"
-	cacheExpirationEnvKey    = "KMS_CACHE_EXPIRATION"
-	cacheExpirationFlagUsage = "An optional value for cache expiration. If not set caching is disabled. Supports " +
-		"valid duration strings, e.g. 10m, 60s, etc. " + commonEnvVarUsageText + cacheExpirationEnvKey
-)
-
-// Hub Auth integration parameters.
-const (
-	hubAuthURLFlagName  = "hub-auth-url"
-	hubAuthURLEnvKey    = "KMS_HUB_AUTH_URL"
-	hubAuthURLFlagUsage = "The URL of Hub Auth server to use for fetching secret share for secret lock. If not " +
-		"specified secret lock based on primary key is used. " + commonEnvVarUsageText + hubAuthURLEnvKey
-
-	hubAuthAPITokenFlagName  = "hub-auth-api-token"     //nolint:gosec // not hard-coded credentials
-	hubAuthAPITokenEnvKey    = "KMS_HUB_AUTH_API_TOKEN" //nolint:gosec // not hard-coded credentials
-	hubAuthAPITokenFlagUsage = "A static token used to protect the GET /secrets API in Hub Auth. " +
-		commonEnvVarUsageText + hubAuthAPITokenEnvKey
-)
-
-const (
-	enableZCAPsFlagName  = "enable-zcaps"
-	enableZCAPsFlagUsage = "Enables ZCAPs authz on all endpoints (except createKeyStore). Default is false. " +
-		commonEnvVarUsageText + enableZCAPsEnvKey
-	enableZCAPsEnvKey = "KMS_ZCAP_ENABLE"
-
-	enableCORSFlagName  = "enable-cors"
-	enableCORSFlagUsage = "Enables CORS. Possible values [true] [false]. " +
-		"Defaults to false if not set. " + commonEnvVarUsageText + corsEnableEnvKey
-	corsEnableEnvKey = "KMS_CORS_ENABLE"
-)
-
-const (
-	syncTimeoutFlagName  = "sync-timeout"
-	syncTimeoutFlagUsage = "Total time in seconds to resolve config values." +
-		" Alternatively, this can be set with the following environment variable: " + syncTimeoutEnvKey
-	syncTimeoutEnvKey = "KMS_SYNC_TIMEOUT"
-)
-
-const (
-	storageTypeMemOption     = "mem"
-	storageTypeCouchDBOption = "couchdb"
-	storageTypeMongoDBOption = "mongodb"
-	storageTypeEDVOption     = "edv"
-)
-
-const (
 	keystorePrimaryKeyURI = "local-lock://keystorekms"
 )
-
-const defaultSyncTimeout = "3"
 
 // Server represents an HTTP server.
 type Server interface {
@@ -294,372 +108,6 @@ func createStartCmd(srv Server) *cobra.Command {
 			return startKmsService(parameters, srv)
 		},
 	}
-}
-
-func createFlags(startCmd *cobra.Command) {
-	startCmd.Flags().StringP(hostURLFlagName, "", "", hostURLFlagUsage)
-	startCmd.Flags().StringP(hostMetricsURLFlagName, "", "", hostMetricsURLFlagUsage)
-	startCmd.Flags().StringP(baseURLFlagName, "", "", baseURLFlagUsage)
-
-	startCmd.Flags().StringP(tlsSystemCertPoolFlagName, tlsSystemCertPoolFlagShorthand, "", tlsSystemCertPoolFlagUsage)
-	startCmd.Flags().StringArrayP(tlsCACertsFlagName, tlsCACertsFlagShorthand, []string{}, tlsCACertsFlagUsage)
-
-	startCmd.Flags().StringP(tlsServeCertPathFlagName, "", "", tlsServeCertPathFlagUsage)
-	startCmd.Flags().StringP(tlsServeKeyPathFlagName, "", "", tlsServeKeyPathFlagUsage)
-
-	startCmd.Flags().StringP(logLevelFlagName, logLevelFlagShorthand, "", logLevelPrefixFlagUsage)
-
-	startCmd.Flags().StringP(databaseTypeFlagName, "", "", databaseTypeFlagUsage)
-	startCmd.Flags().StringP(databaseURLFlagName, "", "", databaseURLFlagUsage)
-	startCmd.Flags().StringP(databasePrefixFlagName, "", "", databasePrefixFlagUsage)
-
-	startCmd.Flags().StringP(secretLockKeyPathFlagName, "", "", secretLockKeyPathFlagUsage)
-
-	startCmd.Flags().StringP(primaryKeyDatabaseTypeFlagName, "", "", primaryKeyDatabaseTypeFlagUsage)
-	startCmd.Flags().StringP(primaryKeyDatabaseURLFlagName, "", "", primaryKeyDatabaseURLFlagUsage)
-	startCmd.Flags().StringP(primaryKeyDatabasePrefixFlagName, "", "", primaryKeyDatabasePrefixFlagUsage)
-
-	startCmd.Flags().StringP(localKMSDatabaseTypeFlagName, "", "", localKMSDatabaseTypeFlagUsage)
-	startCmd.Flags().StringP(localKMSDatabaseURLFlagName, "", "", localKMSDatabaseURLFlagUsage)
-	startCmd.Flags().StringP(localKMSDatabasePrefixFlagName, "", "", localKMSDatabasePrefixFlagUsage)
-
-	startCmd.Flags().StringP(keyManagerStorageTypeFlagName, "", "", keyManagerStorageTypeFlagUsage)
-	startCmd.Flags().StringP(keyManagerStorageURLFlagName, "", "", keyManagerStorageURLFlagUsage)
-	startCmd.Flags().StringP(keyManagerStoragePrefixFlagName, "", "", keyManagerStoragePrefixFlagUsage)
-
-	startCmd.Flags().StringP(cacheExpirationFlagName, "", "", cacheExpirationFlagUsage)
-
-	startCmd.Flags().StringP(hubAuthURLFlagName, "", "", hubAuthURLFlagUsage)
-	startCmd.Flags().StringP(hubAuthAPITokenFlagName, "", "", hubAuthAPITokenFlagUsage)
-
-	startCmd.Flags().StringP(enableZCAPsFlagName, "", "", enableZCAPsFlagUsage)
-	startCmd.Flags().StringP(enableCORSFlagName, "", "", enableCORSFlagUsage)
-
-	startCmd.Flags().StringP(didDomainFlagName, "", "", didDomainFlagUsage)
-
-	startCmd.Flags().String(syncTimeoutFlagName, "", syncTimeoutFlagUsage)
-}
-
-type kmsRestParameters struct {
-	hostURL                 string
-	hostMetricsURL          string
-	baseURL                 string
-	tlsUseSystemCertPool    bool
-	tlsCACerts              []string
-	tlsServeParams          *tlsServeParameters
-	storageParams           *storageParameters
-	secretLockKeyPath       string
-	primaryKeyStorageParams *storageParameters
-	localKMSStorageParams   *storageParameters
-	keyManagerStorageParams *storageParameters
-	cacheExpiration         string
-	hubAuthURL              string
-	hubAuthAPIToken         string
-	logLevel                string
-	enableZCAPs             bool
-	enableCORS              bool
-	didDomain               string
-	syncTimeout             uint64
-}
-
-type tlsServeParameters struct {
-	certPath string
-	keyPath  string
-}
-
-type storageParameters struct {
-	storageType   string
-	storageURL    string
-	storagePrefix string
-}
-
-//nolint:gocyclo // no complicated logic here.
-func getKmsRestParameters(cmd *cobra.Command) (*kmsRestParameters, error) { //nolint:funlen // better readability
-	hostURL, err := cmdutils.GetUserSetVarFromString(cmd, hostURLFlagName, hostURLEnvKey, false)
-	if err != nil {
-		return nil, err
-	}
-
-	hostMetricsURL, err := cmdutils.GetUserSetVarFromString(cmd, hostMetricsURLFlagName, hostMetricsURLEnvKey, true)
-	if err != nil {
-		return nil, err
-	}
-
-	baseURL, err := cmdutils.GetUserSetVarFromString(cmd, baseURLFlagName, baseURLEnvKey, true)
-	if err != nil {
-		return nil, err
-	}
-
-	tlsUseSystemCertPool, tlsCACerts, err := getTLS(cmd)
-	if err != nil {
-		return nil, err
-	}
-
-	tlsServeParams, err := getServeTLS(cmd)
-	if err != nil {
-		return nil, err
-	}
-
-	logLevel, err := cmdutils.GetUserSetVarFromString(cmd, logLevelFlagName, logLevelEnvKey, true)
-	if err != nil {
-		return nil, err
-	}
-
-	storageParams, err := getStorageParameters(cmd)
-	if err != nil {
-		return nil, err
-	}
-
-	secretLockKeyPath, err := cmdutils.GetUserSetVarFromString(cmd, secretLockKeyPathFlagName,
-		secretLockKeyPathEnvKey, true)
-	if err != nil {
-		return nil, err
-	}
-
-	primaryKeyStorageParams, err := getPrimaryKeyStorageParameters(cmd)
-	if err != nil {
-		return nil, err
-	}
-
-	localKMSStorageParams, err := getLocalKMSStorageParameters(cmd)
-	if err != nil {
-		return nil, err
-	}
-
-	keyManagerStorageParams, err := getKeyManagerStorageParameters(cmd)
-	if err != nil {
-		return nil, err
-	}
-
-	cacheExpiration, err := cmdutils.GetUserSetVarFromString(cmd, cacheExpirationFlagName, cacheExpirationEnvKey, true)
-	if err != nil {
-		return nil, err
-	}
-
-	hubAuthURL, err := cmdutils.GetUserSetVarFromString(cmd, hubAuthURLFlagName, hubAuthURLEnvKey, true)
-	if err != nil {
-		return nil, err
-	}
-
-	hubAuthAPIToken, err := cmdutils.GetUserSetVarFromString(cmd, hubAuthAPITokenFlagName, hubAuthAPITokenEnvKey, true)
-	if err != nil {
-		return nil, err
-	}
-
-	didDomain, err := cmdutils.GetUserSetVarFromString(cmd, didDomainFlagName, didDomainEnvKey, true)
-	if err != nil {
-		return nil, err
-	}
-
-	syncTimeoutStr, err := cmdutils.GetUserSetVarFromString(cmd, syncTimeoutFlagName, syncTimeoutEnvKey, true)
-	if err != nil {
-		return nil, err
-	}
-
-	if syncTimeoutStr == "" {
-		syncTimeoutStr = defaultSyncTimeout
-	}
-
-	syncTimeout, err := strconv.ParseUint(syncTimeoutStr, 10, 64)
-	if err != nil {
-		return nil, fmt.Errorf("sync timeout is not a number(positive): %w", err)
-	}
-
-	enableZCAPsConfig, err := cmdutils.GetUserSetVarFromString(cmd, enableZCAPsFlagName, enableZCAPsEnvKey, true)
-	if err != nil {
-		return nil, err
-	}
-
-	enableZCAPs := false
-
-	if enableZCAPsConfig != "" {
-		enableZCAPs, err = strconv.ParseBool(enableZCAPsConfig)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	enableCORS, err := getEnableCORS(cmd)
-	if err != nil {
-		return nil, err
-	}
-
-	return &kmsRestParameters{
-		hostURL:                 strings.TrimSpace(hostURL),
-		hostMetricsURL:          hostMetricsURL,
-		baseURL:                 baseURL,
-		tlsUseSystemCertPool:    tlsUseSystemCertPool,
-		tlsCACerts:              tlsCACerts,
-		tlsServeParams:          tlsServeParams,
-		storageParams:           storageParams,
-		secretLockKeyPath:       secretLockKeyPath,
-		primaryKeyStorageParams: primaryKeyStorageParams,
-		localKMSStorageParams:   localKMSStorageParams,
-		keyManagerStorageParams: keyManagerStorageParams,
-		cacheExpiration:         cacheExpiration,
-		hubAuthURL:              hubAuthURL,
-		hubAuthAPIToken:         hubAuthAPIToken,
-		logLevel:                logLevel,
-		enableZCAPs:             enableZCAPs,
-		enableCORS:              enableCORS,
-		didDomain:               didDomain,
-		syncTimeout:             syncTimeout,
-	}, nil
-}
-
-func getTLS(cmd *cobra.Command) (bool, []string, error) {
-	tlsSystemCertPoolString := cmdutils.GetUserSetOptionalVarFromString(cmd, tlsSystemCertPoolFlagName,
-		tlsSystemCertPoolEnvKey)
-
-	tlsUseSystemCertPool := false
-
-	if tlsSystemCertPoolString != "" {
-		var err error
-		tlsUseSystemCertPool, err = strconv.ParseBool(tlsSystemCertPoolString)
-
-		if err != nil {
-			return false, nil, err
-		}
-	}
-
-	tlsCACerts := cmdutils.GetUserSetOptionalVarFromArrayString(cmd, tlsCACertsFlagName, tlsCACertsEnvKey)
-
-	return tlsUseSystemCertPool, tlsCACerts, nil
-}
-
-func getServeTLS(cmd *cobra.Command) (*tlsServeParameters, error) {
-	tlsCertPath, err := cmdutils.GetUserSetVarFromString(cmd, tlsServeCertPathFlagName, tlsServeCertPathEnvKey, true)
-	if err != nil {
-		return nil, err
-	}
-
-	tlsKeyPath, err := cmdutils.GetUserSetVarFromString(cmd, tlsServeKeyPathFlagName, tlsServeKeyPathFlagEnvKey, true)
-	if err != nil {
-		return nil, err
-	}
-
-	return &tlsServeParameters{
-		certPath: tlsCertPath,
-		keyPath:  tlsKeyPath,
-	}, nil
-}
-
-func getEnableCORS(cmd *cobra.Command) (bool, error) {
-	enableCORSString := cmdutils.GetUserSetOptionalVarFromString(cmd, enableCORSFlagName, corsEnableEnvKey)
-
-	enableCORS := false
-
-	if enableCORSString != "" {
-		var err error
-		enableCORS, err = strconv.ParseBool(enableCORSString)
-
-		if err != nil {
-			return false, err
-		}
-	}
-
-	return enableCORS, nil
-}
-
-func getStorageParameters(cmd *cobra.Command) (*storageParameters, error) {
-	dbType, err := cmdutils.GetUserSetVarFromString(cmd, databaseTypeFlagName, databaseTypeEnvKey, false)
-	if err != nil {
-		return nil, err
-	}
-
-	dbURL, err := cmdutils.GetUserSetVarFromString(cmd, databaseURLFlagName, databaseURLEnvKey, true)
-	if err != nil {
-		return nil, err
-	}
-
-	dbPrefix, err := cmdutils.GetUserSetVarFromString(cmd, databasePrefixFlagName, databasePrefixEnvKey, true)
-	if err != nil {
-		return nil, err
-	}
-
-	return &storageParameters{
-		storageType:   dbType,
-		storageURL:    dbURL,
-		storagePrefix: dbPrefix,
-	}, nil
-}
-
-func getPrimaryKeyStorageParameters(cmd *cobra.Command) (*storageParameters, error) {
-	storageType, err := cmdutils.GetUserSetVarFromString(cmd, primaryKeyDatabaseTypeFlagName,
-		primaryKeyDatabaseTypeEnvKey, false)
-	if err != nil {
-		return nil, err
-	}
-
-	storageURL, err := cmdutils.GetUserSetVarFromString(cmd, primaryKeyDatabaseURLFlagName,
-		primaryKeyDatabaseURLEnvKey, true)
-	if err != nil {
-		return nil, err
-	}
-
-	storagePrefix, err := cmdutils.GetUserSetVarFromString(cmd, primaryKeyDatabasePrefixFlagName,
-		primaryKeyDatabasePrefixEnvKey, true)
-	if err != nil {
-		return nil, err
-	}
-
-	return &storageParameters{
-		storageType:   storageType,
-		storageURL:    storageURL,
-		storagePrefix: storagePrefix,
-	}, nil
-}
-
-func getLocalKMSStorageParameters(cmd *cobra.Command) (*storageParameters, error) {
-	dbType, err := cmdutils.GetUserSetVarFromString(cmd, localKMSDatabaseTypeFlagName,
-		localKMSDatabaseTypeEnvKey, false)
-	if err != nil {
-		return nil, err
-	}
-
-	dbURL, err := cmdutils.GetUserSetVarFromString(cmd, localKMSDatabaseURLFlagName,
-		localKMSDatabaseURLEnvKey, true)
-	if err != nil {
-		return nil, err
-	}
-
-	dbPrefix, err := cmdutils.GetUserSetVarFromString(cmd, localKMSDatabasePrefixFlagName,
-		localKMSDatabasePrefixEnvKey, true)
-	if err != nil {
-		return nil, err
-	}
-
-	return &storageParameters{
-		storageType:   dbType,
-		storageURL:    dbURL,
-		storagePrefix: dbPrefix,
-	}, nil
-}
-
-func getKeyManagerStorageParameters(cmd *cobra.Command) (*storageParameters, error) {
-	storageType, err := cmdutils.GetUserSetVarFromString(cmd, keyManagerStorageTypeFlagName,
-		keyManagerStorageTypeEnvKey, false)
-	if err != nil {
-		return nil, err
-	}
-
-	storageURL, err := cmdutils.GetUserSetVarFromString(cmd, keyManagerStorageURLFlagName,
-		keyManagerStorageURLEnvKey, true)
-	if err != nil {
-		return nil, err
-	}
-
-	storagePrefix, err := cmdutils.GetUserSetVarFromString(cmd, keyManagerStoragePrefixFlagName,
-		keyManagerStoragePrefixEnvKey, true)
-	if err != nil {
-		return nil, err
-	}
-
-	return &storageParameters{
-		storageType:   storageType,
-		storageURL:    storageURL,
-		storagePrefix: storagePrefix,
-	}, nil
 }
 
 func startKmsService(params *kmsRestParameters, srv Server) error {
@@ -763,17 +211,12 @@ func prepareOperationConfig(params *kmsRestParameters) (*operation.Config, error
 		return nil, err
 	}
 
-	primaryKeyStorageProvider, err := prepareKMSStorageProvider(params.primaryKeyStorageParams)
+	primaryKeyLock, err := preparePrimaryKeyLock(storageProvider, params.secretLockKeyPath, params.syncTimeout)
 	if err != nil {
 		return nil, err
 	}
 
-	primaryKeyLock, err := preparePrimaryKeyLock(primaryKeyStorageProvider, params.secretLockKeyPath, params.syncTimeout)
-	if err != nil {
-		return nil, err
-	}
-
-	localKMS, err := prepareLocalKMS(primaryKeyLock, params)
+	localKMS, err := prepareLocalKMS(storageProvider, primaryKeyLock)
 	if err != nil {
 		return nil, err
 	}
@@ -798,8 +241,7 @@ func prepareOperationConfig(params *kmsRestParameters) (*operation.Config, error
 		return nil, err
 	}
 
-	kmsService, err := prepareKMSService(storageProvider, primaryKeyStorageProvider, primaryKeyLock,
-		localKMS, cryptoService, authService, params)
+	kmsService, err := prepareKMSService(storageProvider, primaryKeyLock, localKMS, cryptoService, authService, params)
 	if err != nil {
 		return nil, err
 	}
@@ -890,12 +332,7 @@ func prepareVDR(params *kmsRestParameters) (zcapldcore.VDRResolver, error) {
 	), nil
 }
 
-func prepareLocalKMS(primaryKeyLock secretlock.Service, params *kmsRestParameters) (arieskms.KeyManager, error) {
-	storageProvider, err := prepareKMSStorageProvider(params.localKMSStorageParams)
-	if err != nil {
-		return nil, err
-	}
-
+func prepareLocalKMS(storageProvider storage.Provider, primaryKeyLock secretlock.Service) (arieskms.KeyManager, error) {
 	provider := &kmsProvider{
 		storageProvider: storageProvider,
 		secretLock:      primaryKeyLock,
@@ -905,13 +342,13 @@ func prepareLocalKMS(primaryKeyLock secretlock.Service, params *kmsRestParameter
 }
 
 //nolint:funlen // ignore
-func prepareKMSService(storageProvider, primaryKeyStorageProvider storage.Provider, primaryKeyLock secretlock.Service,
+func prepareKMSService(storageProvider storage.Provider, primaryKeyLock secretlock.Service,
 	localKMS arieskms.KeyManager, cryptoService crypto.Crypto, signer edv.HeaderSigner,
 	params *kmsRestParameters) (kms.Service, error) {
 	var (
-		cacheProvider             storage.Provider
-		keyManagerStorageProvider storage.Provider
-		edvServerURL              string
+		cacheProvider           storage.Provider
+		userKeysStorageProvider storage.Provider
+		edvServerURL            string
 	)
 
 	if params.cacheExpiration != "" {
@@ -923,15 +360,15 @@ func prepareKMSService(storageProvider, primaryKeyStorageProvider storage.Provid
 		cacheProvider = cache.NewProvider(cache.WithExpiration(exp))
 	}
 
-	if params.keyManagerStorageParams.storageType == storageTypeEDVOption {
-		edvServerURL = params.keyManagerStorageParams.storageURL
+	if params.userKeysStorageParams.storageType == storageTypeEDVOption {
+		edvServerURL = params.userKeysStorageParams.storageURL
 	} else {
-		p, err := prepareKMSStorageProvider(params.keyManagerStorageParams)
+		p, err := prepareStorageProvider(params.userKeysStorageParams)
 		if err != nil {
 			return nil, err
 		}
 
-		keyManagerStorageProvider = p
+		userKeysStorageProvider = p
 	}
 
 	rootCAs, err := tlsutils.GetCertPool(params.tlsUseSystemCertPool, params.tlsCACerts)
@@ -948,22 +385,21 @@ func prepareKMSService(storageProvider, primaryKeyStorageProvider storage.Provid
 	}
 
 	config := &kms.Config{
-		StorageProvider:           storageProvider,
-		CacheProvider:             cacheProvider,
-		KeyManagerStorageProvider: keyManagerStorageProvider,
-		LocalKMS:                  localKMS,
-		CryptoService:             cryptoService,
-		HeaderSigner:              signer,
-		PrimaryKeyStorageProvider: primaryKeyStorageProvider,
-		PrimaryKeyLock:            primaryKeyLock,
-		CreateSecretLockFunc:      lock.New,
-		EDVServerURL:              edvServerURL,
-		HubAuthURL:                params.hubAuthURL,
-		HubAuthAPIToken:           params.hubAuthAPIToken,
-		HTTPClient:                httpClient,
-		TLSConfig:                 tlsConfig,
-		SyncTimeout:               params.syncTimeout,
-		Metrics:                   &noopMetrics{},
+		StorageProvider:         storageProvider,
+		CacheProvider:           cacheProvider,
+		UserKeysStorageProvider: userKeysStorageProvider,
+		LocalKMS:                localKMS,
+		CryptoService:           cryptoService,
+		HeaderSigner:            signer,
+		PrimaryKeyLock:          primaryKeyLock,
+		CreateSecretLockFunc:    lock.New,
+		EDVServerURL:            edvServerURL,
+		HubAuthURL:              params.hubAuthURL,
+		HubAuthAPIToken:         params.hubAuthAPIToken,
+		HTTPClient:              httpClient,
+		TLSConfig:               tlsConfig,
+		SyncTimeout:             params.syncTimeout,
+		Metrics:                 &noopMetrics{},
 	}
 
 	if params.hostMetricsURL != "" {
@@ -992,19 +428,6 @@ func prepareStorageProvider(params *storageParameters) (storage.Provider, error)
 		return mongodb.NewProvider(params.storageURL, mongodb.WithDBPrefix(params.storagePrefix))
 	default:
 		return nil, errors.New("database not set to a valid type")
-	}
-}
-
-func prepareKMSStorageProvider(params *storageParameters) (storage.Provider, error) {
-	switch {
-	case strings.EqualFold(params.storageType, storageTypeMemOption):
-		return mem.NewProvider(), nil
-	case strings.EqualFold(params.storageType, storageTypeCouchDBOption):
-		return couchdb.NewProvider(params.storageURL, couchdb.WithDBPrefix(params.storagePrefix))
-	case strings.EqualFold(params.storageType, storageTypeMongoDBOption):
-		return mongodb.NewProvider(params.storageURL, mongodb.WithDBPrefix(params.storagePrefix))
-	default:
-		return nil, errors.New("KMS storage not set to a valid type")
 	}
 }
