+++ test/bdd/fixtures/docker-compose.yml
@@ -32,16 +32,10 @@ services:
       - KMS_TLS_SERVE_KEY=/etc/tls/ec-key.pem
       - KMS_DATABASE_TYPE=mongodb
       - KMS_DATABASE_URL=mongodb://mongodb.example.com:27017
-      - KMS_DATABASE_PREFIX=opskeystore_
+      - KMS_DATABASE_PREFIX=opskms_
       - KMS_SECRET_LOCK_KEY_PATH=/etc/tls/secret-lock.key
-      - KMS_PRIMARY_KEY_DATABASE_TYPE=mongodb
-      - KMS_PRIMARY_KEY_DATABASE_URL=mongodb://mongodb.example.com:27017
-      - KMS_PRIMARY_KEY_DATABASE_PREFIX=opskmspk_
-      - KMS_LOCAL_KMS_DATABASE_TYPE=mongodb
-      - KMS_LOCAL_KMS_DATABASE_URL=mongodb://mongodb.example.com:27017
-      - KMS_LOCAL_KMS_DATABASE_PREFIX=opskmslocal_
-      - KMS_KEY_MANAGER_STORAGE_TYPE=edv
-      - KMS_KEY_MANAGER_STORAGE_URL=https://edv.trustbloc.local:8081
+      - KMS_USER_KEYS_STORAGE_TYPE=edv
+      - KMS_USER_KEYS_STORAGE_URL=https://edv.trustbloc.local:8081
       - KMS_CACHE_EXPIRATION=10m
       - KMS_ZCAP_ENABLE=true
       - KMS_LOG_LEVEL=debug
@@ -71,16 +65,10 @@ services:
       - KMS_TLS_SERVE_KEY=/etc/tls/ec-key.pem
       - KMS_DATABASE_TYPE=mongodb
       - KMS_DATABASE_URL=mongodb://mongodb.example.com:27017
-      - KMS_DATABASE_PREFIX=opskeystore_
+      - KMS_DATABASE_PREFIX=opskms_
       - KMS_SECRET_LOCK_KEY_PATH=/etc/tls/secret-lock.key
-      - KMS_PRIMARY_KEY_DATABASE_TYPE=mongodb
-      - KMS_PRIMARY_KEY_DATABASE_URL=mongodb://mongodb.example.com:27017
-      - KMS_PRIMARY_KEY_DATABASE_PREFIX=opskmspk_
-      - KMS_LOCAL_KMS_DATABASE_TYPE=mongodb
-      - KMS_LOCAL_KMS_DATABASE_URL=mongodb://mongodb.example.com:27017
-      - KMS_LOCAL_KMS_DATABASE_PREFIX=opskmslocal_
-      - KMS_KEY_MANAGER_STORAGE_TYPE=edv
-      - KMS_KEY_MANAGER_STORAGE_URL=https://edv.trustbloc.local:8081
+      - KMS_USER_KEYS_STORAGE_TYPE=edv
+      - KMS_USER_KEYS_STORAGE_URL=https://edv.trustbloc.local:8081
       - KMS_CACHE_EXPIRATION=10m
       - KMS_ZCAP_ENABLE=true
       - KMS_LOG_LEVEL=debug
@@ -109,17 +97,11 @@ services:
       - KMS_TLS_SERVE_KEY=/etc/tls/ec-key.pem
       - KMS_DATABASE_TYPE=mongodb
       - KMS_DATABASE_URL=mongodb://mongodb.example.com:27017
-      - KMS_DATABASE_PREFIX=authzkeystore_
+      - KMS_DATABASE_PREFIX=authzkms_
       - KMS_SECRET_LOCK_KEY_PATH=/etc/tls/secret-lock.key
-      - KMS_PRIMARY_KEY_DATABASE_TYPE=mongodb
-      - KMS_PRIMARY_KEY_DATABASE_URL=mongodb://mongodb.example.com:27017
-      - KMS_PRIMARY_KEY_DATABASE_PREFIX=authzkmspk_
-      - KMS_LOCAL_KMS_DATABASE_TYPE=mongodb
-      - KMS_LOCAL_KMS_DATABASE_URL=mongodb://mongodb.example.com:27017
-      - KMS_LOCAL_KMS_DATABASE_PREFIX=authzkmslocal_
-      - KMS_KEY_MANAGER_STORAGE_TYPE=mongodb
-      - KMS_KEY_MANAGER_STORAGE_URL=mongodb://mongodb.example.com:27017
-      - KMS_KEY_MANAGER_STORAGE_PREFIX=authzkmskm_
+      - KMS_USER_KEYS_STORAGE_TYPE=mongodb
+      - KMS_USER_KEYS_STORAGE_URL=mongodb://mongodb.example.com:27017
+      - KMS_USER_KEYS_MANAGER_STORAGE_PREFIX=authzkmskeys_
       - KMS_HUB_AUTH_URL=https://hub-auth.trustbloc.local:8070
       - KMS_HUB_AUTH_API_TOKEN=test_token
       - KMS_CACHE_EXPIRATION=10m
